add more process to get the connectedness plot
I save the name of the dataframe of volatility because I relized that the names would disappear
during the claculation.

I also add a procedure to turn the connectedness table from matrix to datadframe because I realized
that the claculation in f_connectedness return matrix which is not inputable for f_network.

Then I add the network procedure to calcualte the network structure for networkx and plot the
forceatlas plot