trying to make it more flexible
I add a json file to save the needed information for running this
module so that we don't need to specify those variables in the main
script. If the variable is empty in the Preresquite.json, there
will be an input function to ask the user to enter the needed variables.
or the user can just add those variables directly into Prerequestite.json