I modified the __init__ because I don't know how to use inherit appropriately.
I also finish the divide_dataframe function, which solve the reinput
dataframe and period problem and the period of divided dataframe
lower than period specified problem.

I deleted the self.lag because it will recalculate the lag everytime.

I also finish the calculation of rolling connectedness and know try
to build it as panel data.